ANDROID-13479 renamed button.MisticaButton to button2.Button
