+++ catalog/src/main/res/layout/screen_buttons_catalog.xml
@@ -212,14 +212,14 @@
                 android:enabled="false"
                 android:text="Danger Small Disabled" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
                 android:text="Link"
                 app:style="LINK" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
@@ -228,7 +228,7 @@
                 app:iconTint="@color/text_button_link_selector"
                 app:style="LINK" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:id="@+id/link_progress"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -236,7 +236,7 @@
                 android:text="Link Progress"
                 app:style="LINK" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
@@ -244,7 +244,7 @@
                 android:text="Link Disabled"
                 app:style="LINK" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
@@ -252,7 +252,7 @@
                 app:style="LINK"
                 app:withChevron="true" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
@@ -406,14 +406,14 @@
                 android:enabled="false"
                 android:text="Secondary Small Disabled" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
                 android:text="Link"
                 app:style="LINK_INVERSE" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
@@ -422,7 +422,7 @@
                 app:iconTint="@color/text_button_link_inverse_selector"
                 app:style="LINK_INVERSE" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:id="@+id/link_inverse_progress"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -430,7 +430,7 @@
                 android:text="Link Progress"
                 app:style="LINK_INVERSE" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
@@ -438,7 +438,7 @@
                 android:text="Link Disabled"
                 app:style="LINK_INVERSE" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
@@ -446,7 +446,7 @@
                 app:style="LINK_INVERSE"
                 app:withChevron="true" />
 
-            <com.telefonica.mistica.button.MisticaButton
+            <com.telefonica.mistica.button2.Button
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
